Refactoring code structure by extracting managers into managment folder and keeping query_rewriter and rule_generator logic in the core folder. In thi way, we can export the core folder as a git subtree to support other git repos, such as SqlRewriter.
