+++ management/query_manager.py
@@ -1,7 +1,7 @@
 import sys
 # append the path of the parent directory
 sys.path.append("..")
-from core.data_manager import DataManager
+from management.data_manager import DataManager
 import json
